clean up the Searcher interface, fix tests
